+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- The `--verbosity` option's value now propagated to child processes when tests
+  are run out of process. Previously the default `debug` value would always be
+  used.
+
 ## [0.3.1] - 2023-03-03
 
 ### Fixed